backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
gitpython==3.1.40 